move med_repeater logic to med_reminder.py, clean up listener.py
